OrchidOs support: Introduce AppListPreference for a much cleaner look
Something like HeadsUp Blacklist or Gaming Mode show
an app list to the useidr of those which have been
selected in the preference. This is where there is no
control over the sizing or layout of things. See the
following image:

BEFORE: imgur.com/7muwWZ7.jpg

To address this, we will simply extend the Preference
class and add a layout for it to show things properly.
See the following image:

AFTER:  imgur.com/DzgVOjX.jpg

To utilize this, refer to the following patch:
https://review.aosip.dev/c/AOSIP/platform_packages_apps_OwlsNest/+/17438

Signed-off-by: calebcabob <calphonic@gmail.com>
Change-Id: I4563743b75a6fc305bc77f903f0c3d1391d77973
Signed-off-by: Mr- 0 <oldcrocs21@gmail.com>

Update AppListPreference.java

Change-Id: I3ba12e80ee84260982f1ca239045ed08ca486102